Refactor: relocate
